requirements.txt
sentence-transformers==2.2.2
numpy
streamlit
faiss-cpu
sentence-transformers
nltk
